// src/app/views/pages/apps/patients/model/patient.ts
// export class Patient {
//     firstName!: string;
//     lastName!: string;
//     date!: string;
//     time!: string;
//     address!: string;
//     age!: number;
//     gender!: string;
//     email!: string;
//     contactNumber!: number;
//     description!: string;
//     image!: string;
// }

export interface Patient {
    firstName: string;
    lastName: string;
    date: string;
    time: string;
    address: string;
    age: number;
    gender: string;
    email: string;
    contactNumber: number;
    description: string;
    image: string;
}